MET-736 refactorings after review
